fix issue with degree-test phase
This commit changes P's messages in the degree-test phase to be
random linear combinations of the *unencoded* vector (i.e., comm.coeffs)
instead of the encoded vector (i.e., comm.comm).

This is required for soundness of the testing phase, assuming that
the linear code cannot be decoded efficiently. (The alternative is
for P to send V encoded vector and for V to run the decode procedure
to ensure that the random linear combination is indeed a valid code
word---but we don't want to require a decode procedure.)